requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
tavily-python>=0.3.0
huggingface-hub>=0.23.0
Pillow>=10.0.0
